download book covers

+++ main.py
@@ -1,5 +1,6 @@
 import os
 from pathlib import Path
+from urllib.parse import urljoin, urlsplit, unquote
 
 from bs4 import BeautifulSoup
 from pathvalidate import sanitize_filepath, sanitize_filename
@@ -8,20 +9,20 @@ import requests
 
 def check_for_redirect(response: requests.Response):
     for happend_response in response.history:
-        print(happend_response.status_code)
         if happend_response.status_code>=300:
             raise requests.HTTPError
 
 
-def download_picture(url, filename):
+def download_image(url, filename):
+    """!!! filename must be a valid path to the image file !!!"""
     response = requests.get(url)
     response.raise_for_status()
-    with open(filename, 'wb') as file:
-        file.write(response.content)
+    if not os.path.exists(filename):
+        with open(filename, 'wb') as file:
+            file.write(response.content)
 
 
 def download_book(url, filename):
-    print(url)
     response = requests.get(url)
     check_for_redirect(response)
     response.raise_for_status()
@@ -56,6 +57,7 @@ def download_txt(url, filename, folder='books/'):
 
 
 def download_books_with_title():
+    Path('images').mkdir(parents=True, exist_ok=True)
     for i in range(1,11):
         response = requests.get(f'https://tululu.org/b{i}/')
         response.raise_for_status()
@@ -67,10 +69,15 @@ def download_books_with_title():
             soup = BeautifulSoup(response.text, 'lxml')
             splitted_text = soup.find('h1').text.split('::')
             filename = f'{i}. {splitted_text[0].strip()}'
-            try:
-                download_txt(f'https://tululu.org/txt.php?id={i}',filename)
-            except requests.HTTPError:
-                pass
+            img_path = soup.find('div', class_='bookimage').find('img')['src']
+            img_url = urljoin(response.url, img_path)
+            img_file_name = os.path.basename(unquote(img_path))
+            download_image(img_url, os.path.join('images', img_file_name))
+
+            # try:
+            #     download_txt(f'https://tululu.org/txt.php?id={i}',filename)
+            # except requests.HTTPError:
+            #     pass
 
 
     # image_tag = soup.find('img', class_='attachment-post-image')['src']
